Firebase Notification added, Tutorial screen changes
