+++ app/build.gradle
@@ -100,6 +100,7 @@ dependencies {
     compile 'com.google.firebase:firebase-storage:11.4.2'
     compile 'com.google.firebase:firebase-crash:11.4.2'
     compile 'com.google.firebase:firebase-auth:11.4.2'
+    compile 'com.google.firebase:firebase-messaging:11.4.2'
     compile 'com.firebaseui:firebase-ui-auth:3.1.0'
     compile 'com.google.firebase:firebase-invites:11.4.2'
