--- app/prod/release/output.json
@@ -1,1 +1,0 @@
-[{"outputType":{"type":"APK"},"apkInfo":{"type":"MAIN","splits":[],"versionCode":9},"path":"app-prod-release.apk","properties":{"packageId":"directory.tripin.com.tripindirectory","split":"","minSdkVersion":"19"}}]
