+++ app/src/main/java/directory/tripin/com/tripindirectory/newactivities/TutorialScreensActivity.java
@@ -21,15 +21,15 @@ public class TutorialScreensActivity extends AppIntro2 {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        addSlide(AppIntroFragment.newInstance("Search", "", "Best search result for your enquiry", "", R.drawable.ic_search,
+        addSlide(AppIntroFragment.newInstance("Grow Your Business", "", "Add your business in INL so other can search you in the world", "", R.drawable.agreement,
                 ContextCompat.getColor(this, R.color.colorPrimary),
                 Color.WHITE,
                 Color.WHITE));
-        addSlide(AppIntroFragment.newInstance("Up Vote", "", "Give an up-vote if you like ", "", R.drawable.ic_arrow_upward,
+        addSlide(AppIntroFragment.newInstance("Search", "", "Search other by Route, Company Name and City", "", R.drawable.ic_search_tutorial,
                 ContextCompat.getColor(this, R.color.colorPrimary),
                 Color.WHITE,
                 Color.WHITE));
-        addSlide(AppIntroFragment.newInstance("Down Vote", "", "Give an down-vote if you not like", "", R.drawable.ic_arrow_downward,
+        addSlide(AppIntroFragment.newInstance("Just Speak", "", "No need to type just speak we will show you what you want", "", R.drawable.ic_record_voice_over_black_24dp,
                 ContextCompat.getColor(this, R.color.colorPrimary),
                 Color.WHITE,
                 Color.WHITE));
