--- app/src/main/res/drawable-xhdpi-v9/ic_rv_hookup_black_24dp.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-        android:width="24dp"
-        android:height="24dp"
-        android:viewportWidth="24.0"
-        android:viewportHeight="24.0">
-    <path
-        android:fillColor="#ffffff"
-        android:pathData="M20,17v-6c0,-1.1 -0.9,-2 -2,-2L7,9L7,7l-3,3 3,3v-2h4v3L4,14v3c0,1.1 0.9,2 2,2h2c0,1.66 1.34,3 3,3s3,-1.34 3,-3h8v-2h-2zM11,20c-0.55,0 -1,-0.45 -1,-1s0.45,-1 1,-1 1,0.45 1,1 -0.45,1 -1,1zM18,14h-4v-3h4v3zM17,2v2L9,4v2h8v2l3,-3z"/>
-</vector>
