--- app/src/main/res/drawable-xxhdpi-v11/shape_round_corner_unselect.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <corners android:radius="35dp" />
-    <solid android:color="@android:color/white"/>
-    <padding
-        android:bottom="0dp"
-        android:left="0dp"
-        android:right="0dp"
-        android:top="0dp" />
-
-
-</shape>
