+++ app/src/main/res/drawable/agreement.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="487.4dp"
-    android:height="487.4dp"
+    android:width="128dp"
+    android:height="128dp"
     android:viewportWidth="487.4"
     android:viewportHeight="487.4">
 
     <path
-        android:fillColor="#000000"
+        android:fillColor="#FFF"
         android:pathData="M378.6,107.39c9.9,0,19.6-2.6,28.1-7.7l80.7-47.7v-22.3l-90.8,52.9c-6.9,4-14.7,6.1-22.6,6l-120.2-1.4
 c-6-2.2-10.8-3.6-12.6-4.3c-37.6-8.4-67-5.1-87,0.3c-15,4-31,2.5-45.2-4l-109-50v199.8l11.3,6.2c16.7,9.1,30,23.2,38.3,40.3
 l-6.4,5.2c-17.2,12.5-20.9,36.7-9.6,53.8c6.3,8.3,15.2,13.6,25.4,15.4c1.2,0.1,3,0.2,4.3,0.2c3.7,0.2,7.3-0.2,10.4-0.7
