+++ app/src/main/res/drawable/edit_text_round_corner.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners android:radius="2dp" />
+    <solid android:color="@color/arrow_white"/>
+
+    <padding
+        android:bottom="0dp"
+        android:left="0dp"
+        android:right="0dp"
+        android:top="0dp" />
+    <!--<stroke-->
+        <!--android:width="3dp"-->
+        <!--android:color="#b50015" />-->
+
+</shape>
