+++ app/src/main/res/layout/nav_header_home.xml
@@ -23,7 +23,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:paddingTop="@dimen/nav_header_vertical_spacing"
-        android:text="International Logistics Network"
+        android:text="Indian Logistics Network"
         android:textAppearance="@style/TextAppearance.AppCompat.Body1"
         android:textColor="@color/black"
         android:textStyle="bold"/>
