+++ app/src/main/res/values-v21/styles.xml
@@ -12,7 +12,13 @@
     <style name="AppTheme.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
-        <item name="android:statusBarColor">@color/transparent</item>
+    </style>
+
+    <style name="AppTheme.NoActionBarStatusBarTransparent">
+        <item name="windowActionBar">false</item>
+        <item name="windowNoTitle">true</item>
+        <item name="android:statusBarColor">@android:color/transparent</item>
+
     </style>
 </resources>
