+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#da1818</color>
+    <color name="colorPrimary">#ef463d</color>
     <color name="colorPrimaryDark">#b50015</color>
     <color name="colorAccent">#18110e</color>
